Fix append nil to string bug in javascript injection code
After updating to hoptoad_notifier 2.3.3, using rspec in rails, I got quite a number of errors of "can't convert nil to String" that lead back to javascript_notifier.rb, line 36.
In my tests, the body is sometimes nil. I'm not sure if it happens in production or not, but if it can happen in a spec, it's good enough for me to fix.